xRLAgents/requirements.txt
gymnasium
matplotlib
numpy
opencv_python
Pillow
torch